estudo sobre tratamentos de excessão

+++ src/main/java/com/samuel/libraryapi/service/AutorService.java
@@ -1,25 +1,26 @@
 package com.samuel.libraryapi.service;
 
+import com.samuel.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import com.samuel.libraryapi.model.Autor;
 import com.samuel.libraryapi.repository.AutorRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import com.samuel.libraryapi.repository.LivroRepository;
+import com.samuel.libraryapi.validator.AutorValidator;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
 @Service
+@RequiredArgsConstructor
 public class AutorService {
 
     private final AutorRepository repository;
-
-    @Autowired
-    public AutorService(AutorRepository repository) {
-        this.repository = repository;
-    }
+    private final AutorValidator validator;
+    private final LivroRepository livroRepository;
 
     public Autor salvar(Autor autor) {
+        validator.validar(autor);
         return repository.save(autor);
     }
 
@@ -28,6 +29,10 @@ public class AutorService {
     }
 
     public void deletar(Autor autor) {
+        if (possuiLivro(autor)) {
+            throw new OperacaoNaoPermitidaException(
+                    "Não é permitido excluir um autor que possui livros cadastrados");
+        }
         repository.delete(autor);
     }
 
@@ -48,6 +53,11 @@ public class AutorService {
         if(autor.getId() == null){
             throw new IllegalArgumentException("Para atualizar, é necessário que o autor esteja salvo na base");
         }
+        validator.validar(autor);
         repository.save(autor);
+    }
+
+    public boolean possuiLivro(Autor autor) {
+        return livroRepository.existsByAutor(autor);
     }
 }
